feat: P24-08 Phase5 错误表重写+编译警告修复
- 基于微软官方Trappable Errors+DLL STRINGTABLE交叉验证重写vb6_StdErrorDesc
- 恢复735/744/746为合法VB6错误号(微软确认), 同时保留485/486/487
- 补齐缺失条目: 16/17/96/298/320/442/480/31001-31037等
- 修正451描述为Object not a collection
- 修复4类编译警告: C4010/C5286/C4142/C4312
- x64+x86零警告零错误, 76/76回归测试通过

+++ src/rtl/core/vb6rtl.h
@@ -481,8 +481,8 @@ BSTR vb6_Oct(int32_t n);
 // P18-A: 兼容性填平 — 新增RTL函数
 int64_t vb6_CCur(double v);            // CCur: value * 10000
 vb6_VARIANT vb6_CDec(vb6_VARIANT v);     // P20-07: CDec返回真实DECIMAL (vt=14)
-int32_t vb6_RGB(int32_t r, int32_t g, int32_t b);  // RGB: OLE color
-int32_t vb6_QBColor(int32_t n);        // QBColor: 16-color lookup
+long vb6_RGB(int32_t r, int32_t g, int32_t b);  // RGB: OLE color
+long vb6_QBColor(int32_t n);        // QBColor: 16-color lookup
 double vb6_FileDateTime(BSTR pathname); // FileDateTime: 文件修改时间→VB6 date serial
 int32_t vb6_FileLen(BSTR pathname);    // FileLen: 文件大小(字节)
 void   vb6_SendKeys(BSTR keys, int32_t wait);    // SendKeys: 发送按键
@@ -737,7 +737,7 @@ double vb6_IRR(void* valuesArray, double guess);
 double vb6_MIRR(void* valuesArray, double financeRate, double reinvestRate);
 
 // P20-37: Registry functions (VB6: SaveSetting/GetSetting/DeleteSetting/GetAllSettings)
-// VB6 registry path: HKEY_CURRENT_USER\Software\VB and VBA Program Settings\
+// VB6 registry path: HKEY_CURRENT_USER\Software\VB and VBA Program Settings
 void vb6_SaveSetting(BSTR appName, BSTR section, BSTR key, BSTR setting);
 BSTR vb6_GetSetting(BSTR appName, BSTR section, BSTR key, BSTR default_);
 void vb6_DeleteSetting(BSTR appName, BSTR section, BSTR key);
